chore: refactored a bit so the inter-processing can be reused
